+++ src/tuneuptechnology/client.js
@@ -5,16 +5,17 @@ module.exports = class Client {
         return 'https://app.tuneuptechnology.com/api/'
     }
     static version() {
-        return '1.1.0'
+        return '1.2.0'
     }
 
-    static async response(data, endpoint) {
+    static async make_http_request(data, endpoint) {
         const request = await got.post(endpoint, {
             headers: {
                 'user-agent': `TuneupTechnologyApp/NodeClient/${Client.version()}`
             },
             json: data,
-            responseType: 'json'
+            responseType: 'json',
+            timeout: 10000
         }).catch(console.log);
         return request.body
     }